feat: Refactor architecture for product cache event
